# swagger-spec.json
{"swagger":"2.0","info":{"description":"API Reframax contratos","version":"1.0","title":"Reframax App"},"basePath":"/","tags":[],"schemes":["http"],"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"paths":{"/planejamentos":{"get":{"responses":{"200":{"description":"Lista de Planejamentos retornadas"},"401":{"description":"Acesso Não Autorizado"},"500":{"description":"Erro Interno"}},"tags":["planejamentos"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"post":{"parameters":[{"name":"Planejamento","required":true,"in":"body","schema":{"$ref":"#/definitions/Planejamento"}}],"responses":{"201":{"description":"Lista de Planejamentos sincronizados"},"401":{"description":"Acesso Não Autorizado"},"500":{"description":"Erro Interno"}},"tags":["planejamentos"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/seguranca/login":{"post":{"parameters":[{"name":"Login","required":true,"in":"body","schema":{"$ref":"#/definitions/Login"}}],"responses":{"200":{"description":"Login efetuado com sucesso"},"401":{"description":"Login Não Autorizado"},"500":{"description":"Erro Interno"}},"tags":["seguranca"],"produces":["application/json"],"consumes":["application/json"]}},"/seguranca/infoUsuario":{"get":{"responses":{"200":{"description":"Informações do usuário retornadas com sucesso"},"401":{"description":"Acesso Não Autorizado"},"500":{"description":"Erro Interno"}},"tags":["seguranca"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"Recurso":{"type":"object","properties":{"uuid":{"type":"string"},"codigo":{"type":"string"},"nome":{"type":"string"},"descricao":{"type":"string"},"quantidade":{"type":"string"}},"required":["uuid","codigo","nome","descricao","quantidade"]},"Equipamento":{"type":"object","properties":{"uuid":{"type":"string"},"nome":{"type":"string"},"descricao":{"type":"string"}},"required":["uuid","nome","descricao"]},"Operario":{"type":"object","properties":{"uuid":{"type":"string"},"nome":{"type":"string"},"matricula":{"type":"string"}},"required":["uuid","nome","matricula"]},"Execucao":{"type":"object","properties":{"uuid":{"type":"string"},"de":{"type":"string"},"ate":{"type":"string"}},"required":["uuid","de","ate"]},"Atividade":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"data":{"type":"string"},"status":{"type":"string"},"codigo":{"type":"string"},"duracao":{"type":"string"},"descricao":{"type":"string"},"recursos":{"type":"array","items":{"$ref":"#/definitions/Recurso"}},"equipamentos":{"type":"array","items":{"$ref":"#/definitions/Equipamento"}},"maoDeObraAlocado":{"type":"array","items":{"$ref":"#/definitions/Operario"}},"maoDeObraDisponivel":{"type":"array","items":{"$ref":"#/definitions/Operario"}},"execucoes":{"type":"array","items":{"$ref":"#/definitions/Execucao"}}},"required":["id","nome","data","status","codigo","duracao","descricao","recursos","equipamentos","maoDeObraAlocado","maoDeObraDisponivel","execucoes"]},"Planejamento":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"data":{"type":"string"},"atividades":{"type":"array","items":{"$ref":"#/definitions/Atividade"}}},"required":["id","nome","data","atividades"]},"Login":{"type":"object","properties":{"matricula":{"type":"string"},"senha":{"type":"string"}},"required":["matricula","senha"]}}}